# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Learning Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /zones:
    get:
      operationId: listZones
      tags: [zones]
      summary: List all zones
      responses:
        "200":
          description: List of zones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Zone"
    post:
      operationId: createZone
      tags: [zones]
      summary: Create a new zone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ZoneInput"
      responses:
        "201":
          description: Zone created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Zone"
  /zones/{id}:
    get:
      operationId: getZone
      tags: [zones]
      summary: Get zone by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Zone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Zone"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── COORDINATORS ─────────────────────────────────────────────────────────────
...
  /trainers:
    get:
      operationId: listTrainers
      tags: [trainers]
      summary: List all trainers, optionally filter by zoneId
      parameters:
        - name: zoneId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of trainers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trainer"
    post:
      operationId: createTrainer
      tags: [trainers]
      summary: Create a new trainer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainerInput"
      responses:
        "201":
          description: Trainer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trainer"
  /trainers/{id}:
    get:
      operationId: getTrainer
      tags: [trainers]
      summary: Get trainer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trainer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trainer"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trainers/{id}/stats:
    get:
      operationId: getTrainerStats
      tags: [trainers]
      summary: Get trainer stats (students, courses, completion)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trainer stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainerStats"

  # ── STUDENTS ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # AUTH
...
    Zone:
      type: object
      required: [id, name, description]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
    ZoneInput:
      type: object
      required: [name, description]
      properties:
        name:
          type: string
        description:
          type: string

    # COORDINATOR
...
    Trainer:
      type: object
      required: [id, name, zoneId, zoneName, studentCount]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        zoneId:
          type: integer
        zoneName:
          type: string
        centreName:
          type: ["string", "null"]
        studentCount:
          type: integer
    TrainerInput:
      type: object
      required: [name, zoneId]
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        zoneId:
          type: integer
        centreName:
          type: ["string", "null"]
    TrainerStats:
      type: object
      required: [trainerId, name, studentCount, completionRate]
      properties:
        trainerId:
          type: integer
        name:
          type: string
        studentCount:
          type: integer
        completionRate:
          type: number

    # STUDENT
...